Funding acknowledgement
Hi,

I'm adding the standard funding acknowledgement that should be added to any public code repository in Cloud4all.

By the way, are you aware that two or more consecutive hyphens in a URL trigger a warning in an HTML validator when you use the URL in a web page? It would therefore be nice if you could rename the repo to something that does not have consecutive hypens (`--`) in its name.